requirements.txt
Flask==3.1.3
Flask_Cors==5.0.0
google_api_python_client==2.172.0
google_auth_oauthlib==1.2.2
mistralai==2.4.4
numpy==2.4.4
protobuf==7.34.1
PyPDF2==3.0.1
python-dotenv==1.2.2
python_docx==1.2.0
scikit_learn==1.8.0
spacy==3.8.14
Werkzeug==3.1.8
